// src/app/shared/model/update-master-cust/update-cust-addr-obj.model.ts
export class UpdateCustAddrObj {
    CustAddrId: any;
    CustId: any;
    MrCustAddrTypeCode: any;
    Addr: any;
    AreaCode1: any;
    AreaCode2: any;
    AreaCode3: any;
    AreaCode4: any;
    City: any;
    Zipcode: any;
    SubZipcode: any;
    MrBuildingOwnershipCode: any;
    PhnArea1: any;
    Phn1: any;
    PhnExt1: any;
    PhnArea2: any;
    Phn2: any;
    PhnExt2: any;
    PhnArea3: any;
    Phn3: any;
    PhnExt3: any;
    FaxArea: any;
    Fax: any;
    Notes: any;
    StayLength: any;
    RowVersion: any;
}